Reorganize non-code data files
